restructure data fetching
