--- ui/src/pages/namespace/Explorer/Pages/schema/dataFetching/request/methods.ts
@@ -1,5 +1,0 @@
-import { z } from "zod";
-
-const MutationMethods = ["POST", "PUT", "PATCH", "DELETE"] as const;
-
-export const MutationMethod = z.enum(MutationMethods);
