--- ui/src/pages/namespace/Explorer/Pages/schema/dataFetching/request/requestHeaders.ts
@@ -1,3 +1,0 @@
-import { KeyValue } from "../../primitives/keyValue";
-
-export const RequestHeaders = KeyValue;
