--- ui/src/pages/namespace/Explorer/Pages/schema/dataFetching/request/searchParams.ts
@@ -1,3 +1,0 @@
-import { KeyValue } from "../../primitives/keyValue";
-
-export const SearchParms = KeyValue;
