+++ ui/src/pages/namespace/Explorer/Pages/schema/primitives/id.ts
@@ -1,0 +1,3 @@
+import { z } from "zod";
+
+export const Id = z.string().min(1);
