+++ ui/src/pages/namespace/Explorer/Pages/schema/procedures/mutation.ts
@@ -1,0 +1,19 @@
+import { DynamicString } from "../primitives/dynamicString";
+import { Id } from "../primitives/id";
+import { KeyValue } from "../primitives/keyValue";
+import { z } from "zod";
+
+const MutationMethods = ["POST", "PUT", "PATCH", "DELETE"] as const;
+
+const MutationMethod = z.enum(MutationMethods);
+
+export const Mutation = z.object({
+  id: Id,
+  method: MutationMethod,
+  endpoint: DynamicString,
+  queryParms: KeyValue,
+  requestHeaders: KeyValue,
+  requestBody: KeyValue,
+});
+
+export type MutationType = z.infer<typeof Mutation>;
