+++ ui/src/pages/namespace/Explorer/Pages/schema/procedures/query.ts
@@ -1,11 +1,11 @@
 import { DynamicString } from "../primitives/dynamicString";
-import { SearchParms } from "./request/searchParams";
+import { KeyValue } from "../primitives/keyValue";
 import { z } from "zod";
 
 export const Query = z.object({
   id: z.string().min(1),
   endpoint: DynamicString,
-  searchParms: SearchParms,
+  searchParms: KeyValue,
 });
 
 export type QueryType = z.infer<typeof Query>;
